+++ src/webapp/components/Layout.js
@@ -6,8 +6,8 @@ export default function Layout({ children }) {
   return (
     <>
       <Helmet
-        titleTemplate="Crown Heights Mutual Aid | %s"
-        defaultTitle="Crown Heights Mutual Aid"
+        titleTemplate="Flatbush United | %s"
+        defaultTitle="Flatbush United"
       >
         <html lang="en" />
         <meta charSet="utf-8" />
